ai_analyst_app/requirements.txt
# --- ⚙️ Web Framework & UI ---
streamlit==1.53.0
fastapi[standard]==0.135.1
uvicorn[standard]==0.41.0
click==8.3.1
requests==2.33.0

# --- 🧠 LLM & LangChain Ecosystem ---
langchain==1.2.12
langchain-core==1.2.19
langchain-openai==1.1.11
langchain-anthropic==1.4.0
langchain-google-genai==4.2.1
langgraph==1.1.2
langchain-community==0.4.1
openai==2.26.0
google-genai==1.59.0 

# --- 🕸️ Data Scraping & Parsing ---
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.14.3
httpx==0.28.1
duckduckgo-search==5.3.0

# --- 📊 Database & Data Analysis ---
SQLAlchemy==2.0.45
pandas==2.2.1
plotly==6.6.0
finance-datareader==0.9.110

# --- 🛠️ Environment, Utilities & Monitoring ---
pydantic==2.12.5
pydantic-settings==2.10.1
python-dotenv==1.2.1
langsmith==0.7.20